Clean up the MI debug output and tests
The MI tests are very difficult to implement, because using splines of degree
higher than 0 clobbers the correct evaluation of the entropy for test images
as small as the given ones. Anyway, the splineparzenMI gradient is already
tested, so is would probably be sufficient to add a test that the MI gradient
is the proper result of the splinepartenMI gradient and the image gradient.