Add a run-time test to select the compression/decompression program
The new run-time tests allows to use different programs for the same
type of compression, i.e. the parallel implementations  pigz will be
used instead of gzip and pxz will be used instead of xz (if available).
Also if the (de)compression program corresponding to the given file type
is not available, then an error is thrown.