Remove boost/filesystem includes from the public interface
This commit removes all includes of boost/filesystem from the
public MIA interfaces. As a result the boost-system interface
is no longer pulled in dependend programs removing the need
to link against it. Closes [tickets:#177].

To achieve this a new backend for collecting the search path for
plug-ins was implemented. In a cleanup, also the old boost filesystem
version 2 interface was removed and the dependency on boost moved to
V >= 1.50.