Update CXDBitImage to not use vector<bool>
The std::vector<bool> is a specialized implementation that is
not a real container. For one thing, access to different elemets
is not thread save, since they might be stored in the same
byte/word etc and it the STL implementation that comes with the
STL does not protect concurrent access to these bits.