Correct the dependencies to alliviate a parallel build failure
If more threads are available then targets to be build the old
dependency chain might invoke building of the man page and the
nipype interface at the same time wich may result in running two
instances of building the helper xml files first and as a result a race
condition occurs that sometimes lets the build fail because the
xml helper file is empty. The patch adds the man page target as
dependency to the nipype target to alliviate this.