Add flag whether to consider  background probabilities in local class selection
Ignoreing the background probabilities has the nice effect that blocking is
reduced in areas with much background but also two or more classes. However,
the background is then also segmented into more than one class if it is not
complietely a homogenous area. With the -B flag one can enable or disable
this.