src/3dsegment-local-cmeans: Fix map access error
In the case that the number of slices N is related to the grid size G
like N = k * G +1 the partitioning would try to one one final pass
on an empty subset resulting in a mapping exception that would exit the
program. Add a check that the start of the last partition is not past the
end of the data.